feat(hub): redesign Inbox review experience
