+++ packages/hub-web/src/components/primitives/alert.tsx
@@ -1,34 +1,21 @@
 import * as React from "react"
-import { cva, type VariantProps } from "class-variance-authority"
 
 import { cn } from "@/lib/utils"
+import styles from "./alert.module.css"
 
-const alertVariants = cva(
-  "group/alert relative grid w-full gap-0.5 rounded-lg border px-2.5 py-2 text-left text-sm has-data-[slot=alert-action]:relative has-data-[slot=alert-action]:pr-18 has-[>svg]:grid-cols-[auto_1fr] has-[>svg]:gap-x-2 *:[svg]:row-span-2 *:[svg]:translate-y-0.5 *:[svg]:text-current *:[svg:not([class*='size-'])]:size-4",
-  {
-    variants: {
-      variant: {
-        default: "bg-card text-card-foreground",
-        destructive:
-          "bg-card text-destructive *:data-[slot=alert-description]:text-destructive/90 *:[svg]:text-current",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  }
-)
+type AlertVariant = "default" | "destructive"
 
 function Alert({
   className,
-  variant,
+  variant = "default",
   ...props
-}: React.ComponentProps<"div"> & VariantProps<typeof alertVariants>) {
+}: React.ComponentProps<"div"> & { variant?: AlertVariant }) {
   return (
     <div
+      className={cn(styles.alert, className)}
       data-slot="alert"
+      data-variant={variant}
       role="alert"
-      className={cn(alertVariants({ variant }), className)}
       {...props}
     />
   )
@@ -37,11 +24,8 @@ function Alert({
 function AlertTitle({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
+      className={cn(styles.title, className)}
       data-slot="alert-title"
-      className={cn(
-        "font-medium group-has-[>svg]/alert:col-start-2 [&_a]:underline [&_a]:underline-offset-3 [&_a]:hover:text-foreground",
-        className
-      )}
       {...props}
     />
   )
@@ -53,11 +37,8 @@ function AlertDescription({
 }: React.ComponentProps<"div">) {
   return (
     <div
+      className={cn(styles.description, className)}
       data-slot="alert-description"
-      className={cn(
-        "text-sm text-balance text-muted-foreground md:text-pretty [&_a]:underline [&_a]:underline-offset-3 [&_a]:hover:text-foreground [&_p:not(:last-child)]:mb-4",
-        className
-      )}
       {...props}
     />
   )
@@ -66,8 +47,8 @@ function AlertDescription({
 function AlertAction({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
+      className={cn(styles.action, className)}
       data-slot="alert-action"
-      className={cn("absolute top-2 right-2", className)}
       {...props}
     />
   )
